+++ drivers/accessibility/speakup/main.c
@@ -257,7 +257,7 @@ static struct notifier_block vt_notifier_block = {
 
 static unsigned char get_attributes(struct vc_data *vc, u16 *pos)
 {
-	pos = screen_pos(vc, pos - (u16 *)vc->vc_origin, true);
+	pos = screen_pos(vc, pos - (u16 *)(uintptr_t)vc->vc_origin, true);
 	return (scr_readw(pos) & ~vc->vc_hi_font_mask) >> 8;
 }
 
@@ -462,7 +462,7 @@ static u16 get_char(struct vc_data *vc, u16 *pos, u_char *attribs)
 		u16 w;
 		u16 c;
 
-		pos = screen_pos(vc, pos - (u16 *)vc->vc_origin, true);
+		pos = screen_pos(vc, pos - (u16 *)(uintptr_t)vc->vc_origin, true);
 		w = scr_readw(pos);
 		c = w & 0xff;
