+++ drivers/ata/libahci_platform.c
@@ -715,7 +715,7 @@ int ahci_platform_init_host(struct platform_device *pdev,
 	hpriv->irq = irq;
 
 	/* prepare host */
-	pi.private_data = (void *)(unsigned long)hpriv->flags;
+	pi.private_data = (void *)(uintptr_t)hpriv->flags;
 
 	ahci_save_initial_config(dev, hpriv);
