+++ drivers/uio/uio.c
@@ -708,7 +708,7 @@ static vm_fault_t uio_vma_fault(struct vm_fault *vmf)
 	 */
 	offset = (vmf->pgoff - mi) << PAGE_SHIFT;
 
-	addr = (void *)(unsigned long)idev->info->mem[mi].addr + offset;
+	addr = (void *)(uintptr_t)idev->info->mem[mi].addr + offset;
 	if (idev->info->mem[mi].memtype == UIO_MEM_LOGICAL)
 		page = virt_to_page(addr);
 	else
