+++ drivers/video/fbdev/arkfb.c
@@ -1018,7 +1018,7 @@ static int ark_pci_probe(struct pci_dev *dev, const struct pci_device_id *id)
 
 	pcibios_bus_to_resource(dev->bus, &vga_res, &bus_reg);
 
-	par->state.vgabase = (void __iomem *) (unsigned long) vga_res.start;
+	par->state.vgabase = (void __iomem *) (uintptr_t) vga_res.start;
 
 	/* FIXME get memsize */
 	regval = vga_rseq(par->state.vgabase, 0x10);
