+++ drivers/video/fbdev/s3fb.c
@@ -1251,7 +1251,7 @@ static int s3_pci_probe(struct pci_dev *dev, const struct pci_device_id *id)
 
 	pcibios_bus_to_resource(dev->bus, &vga_res, &bus_reg);
 
-	par->state.vgabase = (void __iomem *) (unsigned long) vga_res.start;
+	par->state.vgabase = (void __iomem *) (uintptr_t) vga_res.start;
 
 	/* Unlock regs */
 	cr38 = vga_rcrt(par->state.vgabase, 0x38);
